% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_functions.R
\name{add2}
\alias{add2}
\title{Sum of two vectors of numbers}
\usage{
add2(x, y, negative = FALSE)
}
\arguments{
\item{x}{a vector of numbers}

\item{y}{a vector of numbers}

\item{negative}{logical, whether to flip the sign of the sum}
}
\value{
the following elements:
\itemize{
\item{the sum of two vectors}
\item{...}
}
}
\description{
This function sums two vectors of numbers and optionally
allows for the negative of the sum to be returned.
}
\details{
This paragraph will start the Details section...

This will be the second paragraph of the Details section...
}
\examples{

add(1:5, 6:10)
add(1:5, 6:10, negative=TRUE)

}
