% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_acc.R
\name{calc_acc}
\alias{calc_acc}
\title{Calculate accuracy of classification}
\usage{
calc_acc(actual, predicted)
}
\arguments{
\item{actual}{a vector or a matrix of actual class labels with no \code{NA} or \code{NaN} values.}

\item{predicted}{a vector or a matrix of predicted labels with no \code{NA} or \code{NaN} values.}
}
\value{
Accuracy, the percentage of correctly classified.
}
\description{
This function calculates accuracy given vectors or matrices of actual class and predicted labels.
}
\details{
The input vectors or matrices need to have the same dimension. 

Any \code{NA} or \code{NaN} value in either input will result in \code{NA} of accuracy.
}
\examples{
## simple numeric example:
x = c(1,1)
y = c(1,2)
calc_acc(x, y)

## works for vectors of different types of object
calc_acc(x, as.factor(y))
calc_acc(x, as.character(y))

}
