% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom_gradient.R
\name{gradient.pom}
\alias{gradient.pom}
\title{Gradient function for proportional odds model}
\usage{
gradient.pom(param, y, X)
}
\arguments{
\item{param}{curent values of the alpha and beta parameters of length (J-1+P).}

\item{y}{a vector of ordered factor responses with J levels.}

\item{x}{a N x P data matrix, with no intercept, where categorical variables need to be coded into indicator variables.}
}
\value{
The gradient of the log likelihood of proportional odds model.
}
\description{
This function returns the gradient of log likelihood of the proportional odds model, allowing for
an arbitrary number of J ordered response categories (J>2).
}
\examples{
Given y, X, and the current values of alpha and beta, the gradient of a proportional odds model 
at a specific interation can be calculated as:
gradient.pom(c(alpha, beta), y, X)

}
