% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom_logistic.R
\name{logistic}
\alias{logistic}
\title{This is the logistic function to transform the linear function into the (0,1) range}
\usage{
logistic(t)
}
\arguments{
\item{t}{a linear function}
}
\value{
the logistic function
}
\description{
This is the logistic function to transform the linear function into the (0,1) range
}
\examples{

logistic( a + X \%*\% beta)


}
