% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom_gradient.R
\name{gradient.pom}
\alias{gradient.pom}
\title{POM gradient}
\usage{
gradient.pom(param, y, X)
}
\arguments{
\item{param}{initial values of the alpha and beta parameters of length (J-1+P)}

\item{y}{a vector of ordered factor responses with J levels}

\item{x}{a N x P data matrix, with no intercept, where categorical variables will have been coded into dummy variables}
}
\value{
the gradient of te log likelihood of proportional odds model
}
\description{
This function returns the gradient of the ordered logit model log likelihood, allowing
an arbitrary number of J ordered response categories (J>2)
}
\examples{

gradient.pom(c(a_vec, beta_vec), y, X)

}
