% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom_logistic.R
\name{logistic}
\alias{logistic}
\title{Standard logistic function}
\usage{
logistic(t)
}
\arguments{
\item{t}{a real number or a real-numbered vector.}
}
\value{
The transformed t.
}
\description{
This is the standard logistic function to transform real numbers into the (0,1) range.
}
\details{
The standard logistic function is given by f(t) = 1 / (1 + exp(-t)).
}
\examples{
For a logistic regression model, the estimated probability of an event associated 
with covariates Xi can be calculated as:
logistic(a + Xi \%*\% beta_hat)


}
