% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom_recode.R
\name{recode.factor}
\alias{recode.factor}
\title{Create a design matrix without intercept}
\usage{
\method{recode}{factor}(formula, data)
}
\arguments{
\item{formula}{a formula object connected by "~", using the column names of the data matrix.}

\item{data}{N x C data matrix without intercept. C is the number of variables.}
}
\value{
The model matrix coded with indicator variables and without intercepts.
}
\description{
This function spans the data matrix according to the user input formula and removes the default intercept, 
as response-specific intercepts are supplied in the \code{loglik.pom} function.
}
\details{
It uses the built-in \code{model.matrix} function so that logical and string variables are converted to indicator variables.
}
\examples{
X2 = recode.factor(Severity_c ~ Source + Traffic_Signal + Sunrise_Sunset + weekday + interstate, data = tst_data)


}
